embodied_ssl/ssl_baselines/rl/ddppo/rotation_reward.py
from typing import Any, List, Optional, Tuple

import attr
import numpy as np
import quaternion
from gym import spaces

import habitat
from habitat.config import Config
from habitat.core.dataset import Dataset, Episode
from habitat.core.embodied_task import (
    EmbodiedTask,
    Measure,
    SimulatorTaskAction,
)
from habitat.core.logging import logger
from habitat.core.registry import registry
from habitat.core.simulator import (
    AgentState,
    RGBSensor,
    Sensor,
    SensorTypes,
    ShortestPathPoint,
    Simulator,
)
from habitat.tasks.nav.nav import NavigationEpisode
from habitat.utils.geometry_utils import quaternion_from_coeff, angle_between_quaternions

from habitat_baselines.common.baseline_registry import baseline_registry


@baseline_registry.register_env(name="NavRLEnvNew")
class NavRLEnvNew(habitat.RLEnv):
    def __init__(self, config: Config, dataset: Optional[Dataset] = None):
        self._rl_config = config.RL
        self._core_env_config = config.TASK_CONFIG
        self._reward_measure_name = self._rl_config.REWARD_MEASURE
        self._success_measure_name = self._rl_config.SUCCESS_MEASURE
        self._angle_measure_name = self._rl_config.ANGLE_MEASURE

        self._previous_measure = None
        self._previous_angle = None
        self._previous_action = None
        super().__init__(self._core_env_config, dataset)

    def reset(self):
        self._previous_action = None
        observations = super().reset()
        self._previous_measure = self._env.get_metrics()[
            self._reward_measure_name
        ]
        self._previous_angle = self._env.get_metrics()[
            self._angle_measure_name
        ]
        return observations

    def step(self, *args, **kwargs):
        self._previous_action = kwargs["action"]
        return super().step(*args, **kwargs)

    def get_reward_range(self):
        return (
            self._rl_config.SLACK_REWARD - 1.0,
            self._rl_config.SUCCESS_REWARD + 1.0,
        )

    def get_reward(self, observations):
        reward = self._rl_config.SLACK_REWARD

        current_measure = self._env.get_metrics()[self._reward_measure_name]

        reward += self._previous_measure - current_measure
        self._previous_measure = current_measure

        # angle reward
        current_angle = self._env.get_metrics()[self._angle_measure_name]
        self._previous_angle = current_angle
        if current_measure < self._core_env_config.TASK.SUCCESS.SUCCESS_DISTANCE:
            reward += self._previous_angle - current_angle

        if self._episode_success():
            reward += self._rl_config.SUCCESS_REWARD
            if current_angle < self._core_env_config.TASK.SUCCESS.SUCCESS_ANGLE:
                reward += self._rl_config.SUCCESS_REWARD

        return reward

    def _episode_success(self):
        return self._env.get_metrics()[self._success_measure_name]

    def get_done(self, observations):
        done = False
        if self._env.episode_over or self._episode_success():
            done = True
        return done

    def get_info(self, observations):
        return self.habitat_env.get_metrics()

@registry.register_measure
class AngleToGoal(Measure):
    """The measure calculates a distance towards the goal."""

    cls_uuid: str = "angle_to_goal"

    def __init__(
        self, sim: Simulator, config: Config, *args: Any, **kwargs: Any
    ):
        self._previous_angle: Optional[float] = None
        self._sim = sim
        self._config = config
        self._episode_view_points: Optional[
            List[float]
        ] = None

        super().__init__(**kwargs)

    def _get_uuid(self, *args: Any, **kwargs: Any) -> str:
        return self.cls_uuid

    def _quat_to_xy_heading(self, quat):
        direction_vector = np.array([0, 0, -1])

        heading_vector = quaternion_rotate_vector(quat, direction_vector)

        phi = cartesian_to_polar(-heading_vector[2], heading_vector[0])[1]
        return np.array([phi], dtype=np.float32)

    def reset_metric(self, episode, *args: Any, **kwargs: Any):
        self._previous_angle = None
        self._metric = None

        self.update_metric(episode=episode, *args, **kwargs)  # type: ignore

    def update_metric(
        self, episode: NavigationEpisode, *args: Any, **kwargs: Any
    ):
        current_angle = self._sim.get_agent_state().rotation

        current_angle_quat = quaternion_from_coeff(current_angle)
        import pdb; pdb.set_trace()

        if self._previous_angle is None or not np.allclose(
            self._previous_angle, current_angle, atol=1e-4
        ):
            goal_angle_quat = quaternion_from_coeff(episode.goals[0].rotation)
            angle_to_target = angle_between_quaternions(current_angle_quat, goal_angle_quat)

            self._previous_angle = current_angle_quat
            self._metric = angle_to_target

@registry.register_sensor
class ImageGoalRotationSensor(Sensor):
    r"""Sensor for ImageGoal observations which are used in ImageGoal Navigation.
    RGBSensor needs to be one of the Simulator sensors.
    This sensor return the rgb image taken from the goal position to reach with
    random rotation.
    Args:
        sim: reference to the simulator for calculating task observations.
        config: config for the ImageGoal sensor.
    """
    cls_uuid: str = "imagegoalrotation"

    def __init__(
        self, *args: Any, sim: Simulator, config: Config, **kwargs: Any
    ):
        self._sim = sim
        sensors = self._sim.sensor_suite.sensors
        rgb_sensor_uuids = [
            uuid
            for uuid, sensor in sensors.items()
            if isinstance(sensor, RGBSensor)
        ]
        if len(rgb_sensor_uuids) != 1:
            raise ValueError(
                f"ImageGoalNav requires one RGB sensor, {len(rgb_sensor_uuids)} detected"
            )

        (self._rgb_sensor_uuid,) = rgb_sensor_uuids
        self._current_episode_id: Optional[str] = None
        self._current_image_goal = None
        super().__init__(config=config)

    def _get_uuid(self, *args: Any, **kwargs: Any) -> str:
        return self.cls_uuid

    def _get_sensor_type(self, *args: Any, **kwargs: Any):
        return SensorTypes.PATH

    def _get_observation_space(self, *args: Any, **kwargs: Any):
        return self._sim.sensor_suite.observation_spaces.spaces[
            self._rgb_sensor_uuid
        ]

    def _get_pointnav_episode_image_goal(self, episode: NavigationEpisode):
        goal_position = np.array(episode.goals[0].position, dtype=np.float32)
        # to be sure that the rotation is the same for the same episode_id
        # since the task is currently using pointnav Dataset.
        # seed = abs(hash(episode.episode_id)) % (2 ** 32)
        # rng = np.random.RandomState(seed)
        angle = rng.uniform(0, 2 * np.pi)
        source_rotation = [0, np.sin(angle / 2), 0, np.cos(angle / 2)]
        # Add source rotation to the 
        episode.goals[0].rotation = source_rotation
        print(f"source rotation: {source_rotation}")
        goal_observation = self._sim.get_observations_at(
            position=goal_position.tolist(), rotation=source_rotation
        )
        return goal_observation[self._rgb_sensor_uuid]

    def get_observation(
        self,
        *args: Any,
        observations,
        episode: NavigationEpisode,
        **kwargs: Any,
    ):
        episode_uniq_id = f"{episode.scene_id} {episode.episode_id}"
        if episode_uniq_id == self._current_episode_id:
            return self._current_image_goal

        self._current_image_goal = self._get_pointnav_episode_image_goal(
            episode
        )
        print("Episode goals location: {}".format(episode.goals[0].rotation))
        self._current_episode_id = episode_uniq_id

        return self._current_image_goal