embodied_ssl/requirements.txt
alembic==1.7.4
antlr4-python3-runtime==4.8
astor==0.8.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.9b0
braceexpand==0.1.7
cfgv==3.3.1
charset-normalizer==2.0.5
click==8.0.1
cloudpickle==2.0.0
configparser==5.0.2
contextlib2==21.6.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.2
filelock==3.0.12
git-archive-all==1.23.0
gitdb==4.0.7
GitPython==3.1.18
gym==0.20.0
hydra-core==1.1.1
identify==2.2.14
idna==3.2
ifcfg==0.22
imageio==2.9.0
imageio-ffmpeg==0.4.5
imgcat==0.5.0
importlib-resources==5.2.2
iniconfig==1.1.1
kiwisolver==1.3.1
kornia==0.6.1
llvmlite==0.36.0
lmdb==1.2.1
lpips==0.1.4
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.3.4
moviepy==2.0.0.dev2
msgpack==1.0.2
mypy-extensions==0.4.3
networkx==2.5.1
nodeenv==1.6.0
numba==0.53.1
numpy==1.19.5
numpy-quaternion==2021.8.30.10.33.11
objectio==0.2.29
omegaconf==2.1.1
opencv-python==4.5.3.56
packaging==21.0
pathspec==0.9.0
pathtools==0.1.2
pickle5==0.0.11
Pillow==8.3.2
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
proglog==0.1.9
promise==2.3
protobuf==3.15.8
psutil==5.8.0
py==1.10.0
pybullet==3.0.4
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
scikit-image==0.17.2
scipy==1.5.4
sentry-sdk==1.4.3
shortuuid==1.0.1
shutils==0.1.0
simplejson==3.17.5
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.3.24
submitit==1.3.3
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==1.15.0
termcolor==1.1.0
tifffile==2020.9.3
timm==0.4.12
toml==0.10.2
tomli==1.2.1
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.2
typed-ast==1.4.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.7.2
visualpriors==0.3.5
wandb==0.12.3
webdataset==0.1.40
yacs==0.1.8
yaspin==2.1.0
